daily

+++ application/migrations/2013_02_26_010938_create_listing.php
@@ -17,7 +17,7 @@ class Create_Listing {
 			$table->text('description');
 			$table->integer('category_id')->unsigned();
 			$table->integer('quantity');
-			$table->float('price');
+			$table->decimal('price', 10, 2);
 			$table->date('date_expiry');
 			$table->date('date_available');
 			$table->date('date_unavailable');
